Start reworking menu to support windows and linux
